refactor: decompose 1100-line dataset_generator.py into domain modules (fixes #2386)
